helper.js clean up
