+++ src/Components/RowBuilder.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, fireEvent, waitFor, screen } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
-import { RowBuilder } from './helpers';
 import { EDIT_MATCH } from '../operations/mutations';
 import '@testing-library/jest-dom';
+import { RowBuilder } from './TableComponents/RowBuilder';
 
 const mutationMock = [
   {
