+++ src/Components/SigDetailsTable/NewSigDetailsTable.js
@@ -27,13 +27,9 @@ import messages from '../../Messages';
 import EmptyAccount from '../SharedComponents/EmptyAccount';
 import './SigDetailsTable.scss';
 import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
-import {
-  BasicTable,
-  filterValues,
-  generateCode,
-  matchStatusFilterMap,
-} from '../helpers';
+import { filterValues, generateCode, matchStatusFilterMap } from '../helpers';
 import { sigDetailsTableColumns } from './Columns';
+import { BasicTable } from '../TableComponents/BasicTable';
 
 const sortIndices = {
   1: 'DISPLAY_NAME',
