+++ src/Components/SysDetailsTable/NewSysDetailsTable.js
@@ -24,13 +24,9 @@ import MessageState from '../MessageState/MessageState';
 import StatusLabel from '../StatusLabel/StatusLabel';
 import messages from '../../Messages';
 import { Icon } from '@patternfly/react-core';
-import {
-  BasicTable,
-  filterValues,
-  generateCode,
-  matchStatusFilterMap,
-} from '../helpers';
+import { filterValues, generateCode, matchStatusFilterMap } from '../helpers';
 import { sysDetailsTableColumns } from './Columns';
+import { BasicTable } from '../TableComponents/BasicTable';
 
 const sortIndices = {
   1: 'NAME',
