+++ src/Components/TableComponents/BasicTable.js
@@ -1,0 +1,39 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Table, Thead, Tr, Th, Tbody } from '@patternfly/react-table';
+import { RowBuilder } from './RowBuilder';
+
+export const BasicTable = ({ hitlistData }) => {
+  const columnNames = {
+    scanDate: 'Dates matched',
+    matchedStatus: 'Match status',
+    note: 'Note',
+  };
+
+  return (
+    <Table
+      aria-label="Simple table"
+      variant="compact"
+      style={{ width: '100%', borderTop: '1px solid #d2d2d2' }}
+    >
+      <Thead style={{ borderBottom: '1px solid #d2d2d2' }}>
+        <Tr>
+          {Object.values(columnNames).map((column) => (
+            <Th key={column}>{column}</Th>
+          ))}
+        </Tr>
+      </Thead>
+      <Tbody>
+        {hitlistData?.map((data) => (
+          <Tr key={data.id}>
+            <RowBuilder data={data} columnNames={columnNames} />
+          </Tr>
+        ))}
+      </Tbody>
+    </Table>
+  );
+};
+
+BasicTable.propTypes = {
+  hitlistData: PropTypes.any,
+};
