+++ src/Components/TableComponents/DropdownBasic.js
@@ -1,0 +1,62 @@
+import React from 'react';
+
+import {
+  Dropdown,
+  DropdownItem,
+  DropdownList,
+  MenuToggle,
+} from '@patternfly/react-core';
+import { useState } from 'react';
+import PropTypes from 'prop-types';
+import { filterValues } from '../helpers';
+
+export const DropdownBasic = ({ dropdownValue, onChange }) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const onToggleClick = () => {
+    setIsOpen(!isOpen);
+  };
+
+  const onSelect = (_event, value) => {
+    onChange(value);
+    setIsOpen(false);
+  };
+
+  return (
+    <Dropdown
+      isOpen={isOpen}
+      onSelect={onSelect}
+      onOpenChange={(isOpen) => setIsOpen(isOpen)}
+      toggle={(toggleRef) => (
+        <MenuToggle
+          aria-label={'MenuToggle'}
+          style={{ maxWidth: '75%' }}
+          ref={toggleRef}
+          onClick={onToggleClick}
+          isExpanded={isOpen}
+        >
+          {filterValues.find((item) => item.value === dropdownValue)?.label}
+        </MenuToggle>
+      )}
+      ouiaId="BasicDropdown"
+    >
+      <DropdownList>
+        {filterValues.map((filterValue) => (
+          <DropdownItem
+            aria-label={`${filterValue.label}`}
+            key={filterValue.value}
+            component="button"
+            onClick={(event) => onSelect(event, filterValue.value)}
+          >
+            {filterValue.label}
+          </DropdownItem>
+        ))}
+      </DropdownList>
+    </Dropdown>
+  );
+};
+
+DropdownBasic.propTypes = {
+  dropdownValue: PropTypes.string,
+  onChange: PropTypes.func,
+};
