+++ src/Components/TableComponents/RowBuilder.js
@@ -1,0 +1,87 @@
+import React from 'react';
+import { Td } from '@patternfly/react-table';
+import PropTypes from 'prop-types';
+import { useCallback, useState } from 'react';
+import { useVerifyInput } from '../useVerifyInput';
+import { useMutation } from '@apollo/client';
+import { EDIT_MATCH } from '../../operations/mutations';
+import { TextInputGroupBasic } from './TextInputBasic';
+import { DropdownBasic } from './DropdownBasic';
+
+function formatDate(inputDate) {
+  const date = new Date(inputDate);
+
+  const day = date.getUTCDate().toString().padStart(2, '0');
+  const month = date.toLocaleString('default', { month: 'long' });
+  const year = date.getUTCFullYear();
+
+  return `${day}-${month}-${year}`;
+}
+
+export const RowBuilder = ({ data, columnNames }) => {
+  const [executeMutation] = useMutation(EDIT_MATCH);
+  const [textValue, setTextValue] = useState(data.justification);
+  const [dropdownValue, setDropdownValue] = useState(data.status);
+
+  // Mutation triggered by dropdown changes
+  const handleDropdownChange = useCallback(
+    (newStatus) => {
+      if (newStatus !== null) {
+        // Check if the newStatus is not null
+        setDropdownValue(newStatus);
+        executeMutation({
+          variables: {
+            input: {
+              sourceHitId: data.id,
+              status: newStatus,
+              justification: textValue, // Use current textValue for consistency
+            },
+          },
+        });
+      }
+    },
+    [executeMutation, data.id]
+  );
+
+  // Debounced execute mutation for text input changes
+  const memoizedExecuteMutation = useCallback(() => {
+    executeMutation({
+      variables: {
+        input: {
+          sourceHitId: data.id,
+          status: dropdownValue,
+          justification: textValue,
+        },
+      },
+    });
+  }, [executeMutation, data.id, textValue]);
+
+  // Hook to perform verification on text input change
+  const verifiedStatus = useVerifyInput(textValue, memoizedExecuteMutation);
+  const isLoading = verifiedStatus === 'loading';
+
+  return (
+    <>
+      <Td dataLabel={columnNames.scanDate}>{formatDate(data.scanDate)}</Td>
+      <Td style={{ maxWidth: '100px' }}>
+        <DropdownBasic
+          dropdownValue={dropdownValue}
+          onChange={handleDropdownChange}
+        />
+      </Td>
+      <Td>
+        <TextInputGroupBasic
+          textValue={textValue}
+          setTextValue={setTextValue}
+          verifiedStatus={verifiedStatus}
+          isLoading={isLoading}
+        />
+      </Td>
+    </>
+  );
+};
+
+RowBuilder.propTypes = {
+  data: PropTypes.object,
+  columnNames: PropTypes.object,
+};
