+++ src/Components/TableComponents/TextInputBasic.js
@@ -1,0 +1,61 @@
+import React from 'react';
+
+import PropTypes from 'prop-types';
+
+import {
+  Flex,
+  HelperText,
+  HelperTextItem,
+  Icon,
+  TextInput,
+  TextInputGroup,
+  TextInputGroupMain,
+} from '@patternfly/react-core';
+
+import { CheckCircleIcon, InProgressIcon } from '@patternfly/react-icons';
+
+export const TextInputGroupBasic = ({
+  textValue,
+  setTextValue,
+  verifiedStatus,
+  isLoading,
+}) => {
+  return (
+    <Flex direction={{ default: 'column' }}>
+      <TextInputGroup>
+        <TextInputGroupMain
+          value={textValue}
+          onChange={(_event, value) => setTextValue(value)}
+        />
+        <TextInput
+          value={textValue}
+          type="text"
+          validated={!isLoading && verifiedStatus === 'error' && verifiedStatus}
+          customIcon={
+            (isLoading && <InProgressIcon />) ||
+            (!isLoading && verifiedStatus === 'success' && (
+              <Icon status="success">
+                <CheckCircleIcon />
+              </Icon>
+            ))
+          }
+          onChange={(_event, value) => setTextValue(value)}
+        />
+      </TextInputGroup>
+      {!isLoading && verifiedStatus === 'error' && (
+        <HelperText>
+          <HelperTextItem variant="error">
+            There was an error saving the note.
+          </HelperTextItem>
+        </HelperText>
+      )}
+    </Flex>
+  );
+};
+
+TextInputGroupBasic.propTypes = {
+  textValue: PropTypes.string,
+  setTextValue: PropTypes.func,
+  verifiedStatus: PropTypes.string,
+  isLoading: PropTypes.string,
+};
