+++ src/Components/helpers.js
@@ -1,24 +1,4 @@
-import React, { useCallback, useState } from 'react';
-import PropTypes from 'prop-types';
-import { Table, Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
-import {
-  Dropdown,
-  DropdownItem,
-  DropdownList,
-  Flex,
-  HelperText,
-  HelperTextItem,
-  Icon,
-  MenuToggle,
-  TextInput,
-  TextInputGroup,
-  TextInputGroupMain,
-} from '@patternfly/react-core';
 import { expandMatchMetadata } from './Common';
-import { useVerifyInput } from './useVerifyInput';
-import { CheckCircleIcon, InProgressIcon } from '@patternfly/react-icons';
-import { EDIT_MATCH } from '../operations/mutations';
-import { useMutation } from '@apollo/client';
 
 export const filterValues = [
   { label: 'Not reviewed', value: 'NOT_REVIEWED' },
@@ -38,216 +18,6 @@ export const matchStatusFilterMap = {
   TEST: 'Malware detection test',
   NO_ACTION: 'No action - risk accepted',
   RESOLVED: 'Resolved - malware removed',
-};
-
-function formatDate(inputDate) {
-  const date = new Date(inputDate);
-
-  const day = date.getUTCDate().toString().padStart(2, '0');
-  const month = date.toLocaleString('default', { month: 'long' });
-  const year = date.getUTCFullYear();
-
-  return `${day}-${month}-${year}`;
-}
-
-export const TextInputGroupBasic = ({
-  textValue,
-  setTextValue,
-  verifiedStatus,
-  isLoading,
-}) => {
-  return (
-    <Flex direction={{ default: 'column' }}>
-      <TextInputGroup>
-        <TextInputGroupMain
-          value={textValue}
-          onChange={(_event, value) => setTextValue(value)}
-        />
-        <TextInput
-          value={textValue}
-          type="text"
-          validated={!isLoading && verifiedStatus === 'error' && verifiedStatus}
-          customIcon={
-            (isLoading && <InProgressIcon />) ||
-            (!isLoading && verifiedStatus === 'success' && (
-              <Icon status="success">
-                <CheckCircleIcon />
-              </Icon>
-            ))
-          }
-          onChange={(_event, value) => setTextValue(value)}
-        />
-      </TextInputGroup>
-      {!isLoading && verifiedStatus === 'error' && (
-        <HelperText>
-          <HelperTextItem variant="error">
-            There was an error saving the note.
-          </HelperTextItem>
-        </HelperText>
-      )}
-    </Flex>
-  );
-};
-
-TextInputGroupBasic.propTypes = {
-  textValue: PropTypes.string,
-  setTextValue: PropTypes.func,
-  verifiedStatus: PropTypes.string,
-  isLoading: PropTypes.string,
-};
-
-export const DropdownBasic = ({ dropdownValue, onChange }) => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const onToggleClick = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const onSelect = (_event, value) => {
-    onChange(value);
-    setIsOpen(false);
-  };
-
-  return (
-    <Dropdown
-      isOpen={isOpen}
-      onSelect={onSelect}
-      onOpenChange={(isOpen) => setIsOpen(isOpen)}
-      toggle={(toggleRef) => (
-        <MenuToggle
-          aria-label={'MenuToggle'}
-          style={{ maxWidth: '75%' }}
-          ref={toggleRef}
-          onClick={onToggleClick}
-          isExpanded={isOpen}
-        >
-          {filterValues.find((item) => item.value === dropdownValue)?.label}
-        </MenuToggle>
-      )}
-      ouiaId="BasicDropdown"
-    >
-      <DropdownList>
-        {filterValues.map((filterValue) => (
-          <DropdownItem
-            aria-label={`${filterValue.label}`}
-            key={filterValue.value}
-            component="button"
-            onClick={(event) => onSelect(event, filterValue.value)}
-          >
-            {filterValue.label}
-          </DropdownItem>
-        ))}
-      </DropdownList>
-    </Dropdown>
-  );
-};
-
-DropdownBasic.propTypes = {
-  dropdownValue: PropTypes.string,
-  onChange: PropTypes.func,
-};
-
-export const RowBuilder = ({ data, columnNames }) => {
-  const [executeMutation] = useMutation(EDIT_MATCH);
-  const [textValue, setTextValue] = useState(data.justification);
-  const [dropdownValue, setDropdownValue] = useState(data.status);
-
-  // Mutation triggered by dropdown changes
-  const handleDropdownChange = useCallback(
-    (newStatus) => {
-      if (newStatus !== null) {
-        // Check if the newStatus is not null
-        setDropdownValue(newStatus);
-        executeMutation({
-          variables: {
-            input: {
-              sourceHitId: data.id,
-              status: newStatus,
-              justification: textValue, // Use current textValue for consistency
-            },
-          },
-        });
-      }
-    },
-    [executeMutation, data.id]
-  );
-
-  // Debounced execute mutation for text input changes
-  const memoizedExecuteMutation = useCallback(() => {
-    executeMutation({
-      variables: {
-        input: {
-          sourceHitId: data.id,
-          status: dropdownValue,
-          justification: textValue,
-        },
-      },
-    });
-  }, [executeMutation, data.id, textValue]);
-
-  // Hook to perform verification on text input change
-  const verifiedStatus = useVerifyInput(textValue, memoizedExecuteMutation);
-  const isLoading = verifiedStatus === 'loading';
-
-  return (
-    <>
-      <Td dataLabel={columnNames.scanDate}>{formatDate(data.scanDate)}</Td>
-      <Td style={{ maxWidth: '100px' }}>
-        <DropdownBasic
-          dropdownValue={dropdownValue}
-          onChange={handleDropdownChange}
-        />
-      </Td>
-      <Td>
-        <TextInputGroupBasic
-          textValue={textValue}
-          setTextValue={setTextValue}
-          verifiedStatus={verifiedStatus}
-          isLoading={isLoading}
-        />
-      </Td>
-    </>
-  );
-};
-
-RowBuilder.propTypes = {
-  data: PropTypes.object,
-  columnNames: PropTypes.object,
-};
-
-export const BasicTable = ({ hitlistData }) => {
-  const columnNames = {
-    scanDate: 'Dates matched',
-    matchedStatus: 'Match status',
-    note: 'Note',
-  };
-
-  return (
-    <Table
-      aria-label="Simple table"
-      variant="compact"
-      style={{ width: '100%', borderTop: '1px solid #d2d2d2' }}
-    >
-      <Thead style={{ borderBottom: '1px solid #d2d2d2' }}>
-        <Tr>
-          {Object.values(columnNames).map((column) => (
-            <Th key={column}>{column}</Th>
-          ))}
-        </Tr>
-      </Thead>
-      <Tbody>
-        {hitlistData?.map((data) => (
-          <Tr key={data.id}>
-            <RowBuilder data={data} columnNames={columnNames} />
-          </Tr>
-        ))}
-      </Tbody>
-    </Table>
-  );
-};
-
-BasicTable.propTypes = {
-  hitlistData: PropTypes.any,
 };
 
 export const generateCode = (host) => {
